requirements.txt
openai==2.14.0
pillow==12.0.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
requests-toolbelt==1.0.0
